added functional tests and reorganized project layout
