+++ .vagrant/machines/default/virtualbox/box_meta
@@ -1,0 +1,1 @@
+{"name":"ubuntu/focal64","version":"20220302.0.0","provider":"virtualbox","directory":"boxes/ubuntu-VAGRANTSLASH-focal64/20220302.0.0/virtualbox"}
