+++ .vagrant/machines/default/virtualbox/creator_uid
@@ -1,0 +1,1 @@
+0
