+++ .vagrant/machines/default/virtualbox/id
@@ -1,0 +1,1 @@
+315c98e9-e1f8-4d3d-b7a6-6db98245157a
