+++ .vagrant/machines/default/virtualbox/index_uuid
@@ -1,0 +1,1 @@
+b927025472014e4988d63908eb31d4e3
