+++ .vagrant/machines/default/virtualbox/private_key
@@ -1,0 +1,27 @@
+-----BEGIN RSA PRIVATE KEY-----
+MIIEowIBAAKCAQEA1QFhQWd9ATDwMeUkrb71vx+ivGvDiS+C+W1Ao132QbOn4wCk
+5asho4LVFdEGMOZ9rVGE9oE8GBdhQX7Ekc/jrxvwh+5oWj4a+Fkr2YNJPfR1hBk/
+dS8yvS95NLKz7Yz/txAAMf3KE8Xz2VOq+5P3YScEnB4vMR24lC2KcEO/Ouu+Rk9o
+5TU2wm7wmU0V9ERaf4zj/YwQ7aHUtBLn4wgJ97e7zDzyFW7aQcKkiaS3m3Snx71/
+vRS9ExQYasiGOp5fHvwtzBrfhlsqFLaYLMd7q/ipE+5bJF+JGTxQaQ5sByOnjUrm
+aazPb2X4/452t7cxoTD+60w9MIdqoVYfToxDbQIDAQABAoIBACcg7MO+Bgmc+tK/
+nEBKjww//H4OccoWr0GwnyojjaQ+I34g1SnXiQNnektarX94J+gZzIoX7Xkjels0
+tVlvzgyR3lqqIgNJ21x5zQNVCbRW3uuYU77sH67/NI3IsiLWPvda+h4019VQDN6E
+bylt4gi3yTIAnXyxUAHwZ7j9yPhClkq9Uonvb3Dm1DiJ5s6VWPES4Me0bYQ9oJzl
+qExRmSceeTR78S8o33HdGFH3QVxWJbG8IBtMyFmxnlMDZo6aJzZDXMkCJq3RIDya
+3Oc6Di6qXNrZ317uqclKE4ICCNTZafmVJ87DwGp7ZKS0Ldo4Y0tqbbu3i8hRSHan
+PX3WlmECgYEA8CJPYeyT36egqSm/8Htb2an7bP21V536CA8DSsjeyAj2LuRmsG++
+8Q/GKeXOHEHMQ6JubBPpmysuubGc4kgfJxBZDIHezapKzg39x9AK4fegtP4dW1oI
+Gw5cZKxLhqyxMGC+Av8XEhvnGkrCxvD7brQfTBwmEYZOKR813xqA8ukCgYEA4xQ1
+gDbfBQ9nBHIURe61b8rrLoS8Wf2dcmNkLWbXl8hGmAGmCUYKGdUHYjo06B9s7452
+3ZOsNragn1T1QXOI0Ca9F7aTxSOUY03plU9BjpFlFWxIVExvcXIdjIOtNbaZ3zLc
+8JoTyj/1M+N+48JDSeTuSlW+5p92ap0IEf6VkeUCgYAd4Ibz7hVv7FFG+tTQvOhu
+qoGHzAE2B4fGXOfR9DATv1mx9+T26tBc+dNz/w2n4l/hwmHuvIOCNNs5a2nTlZma
+qKQLtFTeTePEw2Th6XaRKtgp7LwGU6tSgnToaa9z8CSTmlXJ1acvL52PXeHiErWV
+zn9eXto1r/olj6aKPafxkQKBgQCzCODxuNF3dOBdIfCX/o/wl0V7xM2YChyT7ocy
+QK5kjlQSTuJw11bpbQtIR+okxXonGeTf6ClUbgMP/HTt6MCKlLBrvNifCuFYtjof
+LYDbfgj7rgcM58LJ6BYiDkOHYPPKhZiq3mASCwExqwEFY8mXvjaeRyKfRESh5DNT
+tKyKfQKBgB00r0ic8BaUVzCz55P/jD7hxgOSVnslL1septJhlyocuJ/lI64kjaEm
+ibjJMUtn8tM4Oc3DJcmoYdG0unTrvWlpoC/C0UdzXrWLoORbgBTpXbkw4o4jAMZD
++0zUc0nKc+MRse9+KXK0ScttD07Ugr66OZSqE+g7r0+d/ezfvUmw
+-----END RSA PRIVATE KEY-----
