+++ .vagrant/machines/default/virtualbox/synced_folders
@@ -1,0 +1,1 @@
+{"virtualbox":{"/vagrant":{"guestpath":"/vagrant","hostpath":"C:/Users/tomli/Documents/GitHub/design_project","disabled":false,"__vagrantfile":true}}}
