+++ .vagrant/machines/default/virtualbox/vagrant_cwd
@@ -1,0 +1,1 @@
+C:/Users/tomli/Documents/GitHub/design_project
