+++ src/webapp/flaskr/mealplan.py
@@ -1,7 +1,7 @@
 import os
 import sys
 import json
-from webapp import data_src
+from flaskr import data_src
 import itertools
 import random
