+++ src/webapp/flaskr/templates/exerciseplan.html
@@ -1,0 +1,85 @@
+{% extends 'navigation.html' %}
+
+{% block content %}
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" >
+    <title>Weekly Exercises</title>
+  </head>
+  <style>
+    body { background-image: url("/static/sports.jpg");background-size:cover;background-repeat:no-repeat;width:100% }
+ 
+ 
+  </style>
+  <body>
+
+    <div class="container-fluid" style="margin-top:10px;width:70%">
+      <div>
+        <div class="card">
+            <h5 class="card-header" style="background-color:black;color:white;">Sunday Exercises</h5>
+            <div class="card-body">
+              <h5 class="card-title">Exercise 1</h5>
+              <h5 class="card-title">Exercise 6</h5>
+
+            </div>
+          </div>
+          <div class="card" style="margin-top:10px;">
+            <h5 class="card-header" style="background-color:black;color:white;">Monday Exercises</h5>
+            <div class="card-body">
+              <h5 class="card-title">Exercise 1</h5>
+              <h5 class="card-title">Exercise 6</h5>
+
+            </div>
+          </div>
+          <div class="card" style="margin-top:10px;">
+            <h5 class="card-header" style="background-color:black;color:white;">Tuesday Exercises</h5>
+            <div class="card-body">
+              <h5 class="card-title">Exercise 1</h5>
+              <h5 class="card-title">Exercise 6</h5>
+
+            </div>
+          </div>
+          <div class="card" style="margin-top:10px;">
+            <h5 class="card-header" style="background-color:black;color:white;">Wednesday Exercises</h5>
+            <div class="card-body">
+              <h5 class="card-title">Exercise 1</h5>
+              <h5 class="card-title">Exercise 6</h5>
+
+            </div>
+          </div>
+          <div class="card" style="margin-top:10px;">
+            <h5 class="card-header" style="background-color:black;color:white;">Thursday Exercises</h5>
+            <div class="card-body">
+              <h5 class="card-title">Exercise 1</h5>
+              <h5 class="card-title">Exercise 6</h5>
+
+            </div>
+          </div>
+          <div class="card" style="margin-top:10px;">
+            <h5 class="card-header" style="background-color:black;color:white;">Friday Exercises</h5>
+            <div class="card-body">
+              <h5 class="card-title">Exercise 1</h5>
+              <h5 class="card-title">Exercise 6</h5>
+
+            </div>
+          </div>
+          <div class="card" style="margin-top:10px;">
+            <h5 class="card-header" style="background-color:black;color:white;">Saturday Exercises</h5>
+            <div class="card-body">
+              <h5 class="card-title">Exercise 1</h5>
+              <h5 class="card-title">Exercise 6</h5>
+
+            </div>
+          </div>
+        
+
+      </div>
+    </div>
+
+  </body>
+</html>
+
+{% endblock %}
