+++ src/webapp/flaskr/templates/mealplans.html
@@ -1,0 +1,22 @@
+{% extends 'navigation.html' %}
+
+{% block content %}
+<style>
+    body { background-image: url("/static/foods7.jpg");background-size:cover;background-repeat:no-repeat } 
+  
+  
+  </style>
+    <h1 style="margin-top:10px;text-align:center;background-color:grey;color:black">{% block title %} Personal Meal Plan Recommendations{% endblock %}</h1>
+    <div style="width: 50%; margin: auto">
+        
+        <div style="padding: 15px; background-color: white; margin: 25px">
+            <p style="font-size: 24px">{{ bestmealplan }}</p>
+        </div>
+        {% for meal in mealplans %}
+        <div style="padding: 15px; background-color:white; margin: 25px">
+            <p style="font-size: 24px">{{ meal }}</p>
+        </div>
+{% endfor %}
+        
+    </div>
+{% endblock %}
