+++ src/webapp/flaskr/templates/shoppinglist.html
@@ -1,0 +1,77 @@
+{% extends 'navigation.html' %}
+
+{% block content %}
+<head>
+    <title>Health Planner</title>
+    <link rel="stylesheet" media="screen" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css">
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap-theme.min.css">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+
+ </head>
+ <style>
+    body { background-image: url("/static/foods3.jpg");background-size:cover;background-repeat:no-repeat } 
+
+
+</style>
+
+   <br>
+   <div class="row align-items-center justify-content-center">
+   <div class="col-md-4" style="background-color:grey;color:black;border-radius:10px;width:80%">
+     <form  action="" method="post" role="form">
+       {{ form.csrf }}
+       <h1 style="text-align:center;background-color:grey;color:black">Enter your foods for your personal shopping list</h1>
+
+       <div class="form-group">
+
+         <label for="newfood">Enter new food:</label>
+         <input type="text" class="form-control" id="newfood" name="newfood" placeholder="newfood?">
+          </div>
+       <br>													   
+
+       <center>
+         <button type="submit" class="btn btn-success">Submit</button>
+       </center>
+     </form>
+
+     <br>
+     <div class="list-group">
+      <div class="list-group-item ">
+        <div>
+          <h5 class="mb-1">Foods</h5>
+        </div>
+        <p class="mb-1">Recipe</p>
+        </div>
+        <div class="list-group-item ">
+         <div>
+           <h5 class="mb-1">Foods</h5>
+         </div>
+         <p class="mb-1">Recipe</p>
+         </div>
+         <div class="list-group-item ">
+            <div>
+              <h5 class="mb-1">Foods</h5>
+            </div>
+            <p class="mb-1">Recipe</p>
+            </div>
+            <div class="list-group-item">
+               <div>
+                 <h5 class="mb-1">Foods</h5>
+               </div>
+               <p class="mb-1">Recipe</p>
+               </div>
+               <div class="list-group-item ">
+                  <div>
+                    <h5 class="mb-1">Foods</h5>
+                  </div>
+                  <p class="mb-1">Recipe</p>
+                  </div>
+                  <div class="list-group-item ">
+                     <div>
+                       <h5 class="mb-1">Foods</h5>
+                     </div>
+                     <p class="mb-1">Recipe of the food</p>
+                     </div>
+    </div>
+   </div>
+ 
+{% endblock %}
