--- src/webapp/templates/navigation.html
@@ -1,54 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.0/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-KyZXEAg3QhqLMpG8r+8fhAXLRk2vvoC2f3B09zVXn8CA5QIVfZOJ3BCsw2P0p/We" crossorigin="anonymous">
-
-    <title>{% block title %} {% endblock %} Health Planner</title>
-  </head>
-  <body>
-    <nav class="navbar navbar-expand-lg navbar-light bg-dark">
-    <div class="container-fluid" >
-        <a class="navbar-brand" style="color:white;" href="{{ url_for('login') }}">Health/Diet Planner App</a>
-        <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
-        <span class="navbar-toggler-icon"></span>
-        </button>
-        <div class="collapse navbar-collapse" id="navbarNav">
-        <ul class="navbar-nav">
-          <li class="nav-item">
-            <a class="nav-link" style="color:white;" href="{{ url_for('signup') }}">Sign Up</a>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link" style="color:white;" href="{{ url_for('diet') }}">Meal Planner </a>
-          </li>
-            <li class="nav-item">
-              <a class="nav-link" style="color:white;" href="{{ url_for('exercises') }}">Exercises Form </a>
-            </li>
-            <li class="nav-item">
-              <a class="nav-link" style="color:white;" href="{{ url_for('mealplan') }}">Saved Meal Plan</a>
-            </li>
-            <li class="nav-item">
-              <a class="nav-link" style="color:white;" href="{{ url_for('exerciseplan') }}">Weekly Exercise Plan</a>
-            </li>
-            <li class="nav-item">
-              <a class="nav-link" style="color:white;" href="{{ url_for('listitems') }}">Shopping/Recipes List</a>
-            </li>
-            <li class="nav-item">
-              <a class="nav-link" style="color:white;" href="{{ url_for('points') }}">Your Points</a>
-            </li>
-        </ul>
-        </div>
-    </div>
-    </nav>
-    <div class="container">
-        {% block content %} {% endblock %}
-    </div>
-
-    <!-- Optional JavaScript -->
-
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.0/dist/js/bootstrap.bundle.min.js" integrity="sha384-U1DAWAznBHeqEIlVSCgzq+c9gqGAJn5c/t99JyeKa9xxaYpSvHU5awsuZVVFIhvj" crossorigin="anonymous"></script>
-
-  </body>
-</html>
