--- src/webapp/templates/points.html
@@ -1,88 +1,0 @@
-{% extends "navigation.html" %}
-
-{% block content %}
-<html>
-    <style>
-        body { background-image: url("/static/foods2.jpg");background-size:cover;background-repeat:no-repeat } 
-      
-      
-      </style>
-<body>
-    </h2>
-    <br> 
-    <table style="background-color:white;margin:auto;" cellspacing="30" cellpadding="30">
-        <thead>
-            <h2 style="text-align:center;background-color:black;color: white;width:50%;margin:auto;">Points History Calendar
-
-            <tr>
-                <th>Sun</th>
-                <th>Mon</th>
-                <th>Tue</th>
-                <th>Wed</th>
-                <th>Thu</th>
-                <th>Fri</th>
-                <th>sat</th>
-            </tr>
-        </thead>
-          
-        <tbody>
-            <tr>
-                <td></td>
-                <td></td>
-                <td></td>
-                <td></td>
-                <td></td>
-                <td>Day 1 points: <p>6 points</p></td>
-                <td>Day 2 points: <p>6 points</p></td>
-            </tr>
-            <tr></tr>
-            <tr>
-                <td>Day 3 points: <p>6 points</p></td>
-                <td>Day 4 points: <p>6 points</p></td>
-                <td>Day 5 points: <p>6 points</p></td>
-                <td>Day 6 points: <p>6 points</p></td>
-                <td>7</td>
-                <td>8</td>
-                <td>9</td>
-            </tr>
-            <tr>
-                <td>10</td>
-                <td>11</td>
-                <td>12</td>
-                <td>13</td>
-                <td>14</td>
-                <td>15</td>
-                <td>16</td>
-            </tr>
-            <tr>
-                <td>17</td>
-                <td>18</td>
-                <td>19</td>
-                <td>20</td>
-                <td>21</td>
-                <td>22</td>
-                <td>23</td>
-            </tr>
-            <tr>
-                <td>24</td>
-                <td>25</td>
-                <td>26</td>
-                <td>27</td>
-                <td>28</td>
-                <td>29</td>
-                <td>30</td>
-            </tr>
-            <tr>
-                <td>31</td>
-                <td>1</td>
-                <td>2</td>
-                <td>3</td>
-                <td>4</td>
-                <td>5</td>
-                <td>6</td>
-            </tr>
-        </tbody>
-    </table>
-</body>
-</html>
-{% endblock %}
