--- src/webapp/templates/signup.html
@@ -1,35 +1,0 @@
-{% extends "navigation.html" %}
-
-{% block content %}
-<style>
-  body { background-image: url("/static/foods6.jpg");background-size:cover;background-repeat:no-repeat } 
-
-
-</style>
-<br>
-    <div class="row align-items-center justify-content-center" >
-
-    <div class="col-md-4" style="background-color:grey;color:black;border-radius:10px;width:50%">
-      <form  action="" method="post" role="form">
-        <h1 style="text-align:center;background-color:grey;color:black">Sign Up for a New Health/Diet Planner Account</h1>
-
-        <div class="form-group">
-            <label for="email">Enter email:</label>
-            <input type="email" class="form-control" id="email" name="email" placeholder="email?">
-          <br>  
-          <label for="username">Enter a Username:</label>
-          <input type="text" class="form-control" id="username" name="username" placeholder="username?">
-           </div>
-        <br>	
-        <label for="password">Enter a Password:</label>
-          <input type="text" class="form-control" id="password" name="password" placeholder="password?">
-           												   
-        <center>
-          <button type="submit" class="btn btn-success">Sign Up for a new Account</button>
-        </center>
-      </form>
- 
-      <br>
- 
-    </div>
-    {% endblock %}
